Proposal: Automatically disable autocompletion for models that reject Shout styling (experimental)
Reasoning: Both Autocompletion and Shout follow two relatively similar goals, i.e., to enhance the editing support in arbitrary tools that display source code. By using a simple heuristic that adopts the decision of a tool whether its contents should be syntax-highlighted, we can reduce the adaptation effort of tools for Autocompletion. Since Shout is already a part of the base system, a vast majority of tools can be assumed to implement proper information for syntax highlighting.

Implementation: In StringHolder >> #wantsAutocompletion, redirect to #aboutToStyle:requestor: and pass a pseudo-styler (MessageCatcher) and a nil requestor. For the unlikely event that anything breaks (because some tool might expected a non-nil requestor or a certain styler type), catch any errors, log them to the transcript, and reject autocompletion.

Considerations: #aboutToStyle:requestor: (with the second argument) is a relatively new selector that is only included in Squeak 5.3 and newer. If Autocompletion intends to provide further backward compatibility, a version switch could be introduced. For the pseudo-styler, a more reliable alternative to using a generic MessageCatcher could be instantiating a real SHStylerST80. However, this would add a hard dependency on Shout whereas at the moment, we only rely on the existence of the #aboutToStyle:requestor: protocol which is part of the Kernel package.